chore: cleanup and move mobile app banner
- move mobile app banner to keep everything inline
