+++ sample-apps/react/react-video-demo/src/components/Layout/LobbyLayout/LobbyLayout.tsx
@@ -3,11 +3,10 @@ import classnames from 'classnames';
 import { isAndroid, isIOS, isSafari } from 'mobile-device-detect';
 import { FeatureCollection, Geometry } from 'geojson';
 
+import MobileAppBanner from '../../MobileAppBanner';
 import LatencyMap from '../../LatencyMap';
 
 import styles from './LobbyLayout.module.css';
-import { MobileAppBanner } from '../../Header/MobileAppBanner';
-import { useUserContext } from '../../../contexts/UserContext';
 
 export type Props = {
   className?: string;
@@ -22,8 +21,7 @@ export const LobbyLayout: FC<Props> = ({
   children,
   edges,
 }) => {
-  const { qr } = useUserContext();
-  const shouldRenderMobileAppBanner = qr && (isAndroid || (isIOS && !isSafari));
+  const shouldRenderMobileAppBanner = isAndroid || (isIOS && !isSafari);
   const [isNativeAppsBannerDismissed, setIsNativeAppsBannerDismissed] =
     useState(!shouldRenderMobileAppBanner);
   const rootClassName = classnames(
