+++ sample-apps/react/react-video-demo/src/components/MobileAppBanner/MobileAppBanner.module.css
